requirements.txt
python-frontmatter
sendgrid
PyGithub
markdown2
email_validator
configargparse

setuptools
